refactor(backend-db): update package file structure
