+++ packages/backend/libraries/backend-db/src/persistence/adapter/typeorm/services/FindTypeOrmService.ts
@@ -1,16 +1,18 @@
 import { Builder, BuilderAsync } from '@cornie-js/backend-common';
 import {
   FindManyOptions,
+  FindOneOptions,
   ObjectLiteral,
   QueryBuilder,
   Repository,
   SelectQueryBuilder,
-  WhereExpressionBuilder,
 } from 'typeorm';
 
-import { FindQueryTypeOrmFromQueryWithQueryBuilderBuilder } from '../../builders/typeorm/FindQueryTypeOrmFromQueryWithQueryBuilderBuilder';
+import { FindQueryTypeOrmFromQueryBuilder } from '../builders/FindQueryTypeOrmFromQueryBuilder';
+import { FindQueryTypeOrmFromQueryWithQueryBuilderBuilder } from '../builders/FindQueryTypeOrmFromQueryWithQueryBuilderBuilder';
+import { isQueryBuilder } from '../utils/isQueryBuilder';
 
-export class FindTypeOrmQueryBuilderService<
+export class FindTypeOrmService<
   TModel,
   TModelDb extends ObjectLiteral,
   TQuery,
@@ -19,30 +21,20 @@ export class FindTypeOrmQueryBuilderService<
   readonly #modelFromModelDbBuilder:
     | Builder<TModel, [TModelDb]>
     | BuilderAsync<TModel, [TModelDb]>;
-  readonly #findQueryTypeOrmFromQueryBuilder:
-    | Builder<
-        QueryBuilder<ObjectLiteral> & WhereExpressionBuilder,
-        [TQuery, QueryBuilder<ObjectLiteral> & WhereExpressionBuilder]
-      >
-    | BuilderAsync<
-        QueryBuilder<ObjectLiteral> & WhereExpressionBuilder,
-        [TQuery, QueryBuilder<ObjectLiteral> & WhereExpressionBuilder]
-      >;
+  readonly #findQueryTypeOrmFromQueryBuilder: FindQueryTypeOrmFromQueryBuilder<
+    TModelDb,
+    TQuery
+  >;
 
   constructor(
     repository: Repository<TModelDb>,
     modelFromModelDbBuilder:
       | Builder<TModel, [TModelDb]>
       | BuilderAsync<TModel, [TModelDb]>,
-    findQueryTypeOrmFromQueryBuilder:
-      | Builder<
-          QueryBuilder<ObjectLiteral> & WhereExpressionBuilder,
-          [TQuery, QueryBuilder<ObjectLiteral> & WhereExpressionBuilder]
-        >
-      | BuilderAsync<
-          QueryBuilder<ObjectLiteral> & WhereExpressionBuilder,
-          [TQuery, QueryBuilder<ObjectLiteral> & WhereExpressionBuilder]
-        >,
+    findQueryTypeOrmFromQueryBuilder: FindQueryTypeOrmFromQueryBuilder<
+      TModelDb,
+      TQuery
+    >,
   ) {
     this.#repository = repository;
     this.#modelFromModelDbBuilder = modelFromModelDbBuilder;
@@ -54,6 +46,8 @@ export class FindTypeOrmQueryBuilderService<
       query,
       async (queryBuilder: SelectQueryBuilder<TModelDb>): Promise<TModelDb[]> =>
         queryBuilder.getMany(),
+      async (findManyOptions: FindManyOptions<TModelDb>): Promise<TModelDb[]> =>
+        this.#repository.find(findManyOptions),
     );
 
     const models: TModel[] = await Promise.all(
@@ -72,6 +66,12 @@ export class FindTypeOrmQueryBuilderService<
         queryBuilder: SelectQueryBuilder<TModelDb>,
       ): Promise<TModelDb | undefined> =>
         (await queryBuilder.getOne()) ?? undefined,
+      async (
+        findConditions: FindManyOptions<TModelDb>,
+      ): Promise<TModelDb | undefined> =>
+        (await this.#repository.findOne(
+          this.#buildFindOneOptions(findConditions),
+        )) ?? undefined,
     );
 
     let model: TModel | undefined;
@@ -85,10 +85,19 @@ export class FindTypeOrmQueryBuilderService<
     return model;
   }
 
+  #buildFindOneOptions(
+    findManyOptions: FindManyOptions<TModelDb>,
+  ): FindOneOptions<TModelDb> {
+    return findManyOptions;
+  }
+
   async #innerFind<TOutputDb extends undefined | TModelDb | TModelDb[]>(
     query: TQuery,
     findByQueryBuilder: (
       queryBuilder: SelectQueryBuilder<TModelDb>,
+    ) => Promise<TOutputDb>,
+    findByFindManyOptions: (
+      findManyOptions: FindManyOptions<TModelDb>,
     ) => Promise<TOutputDb>,
   ): Promise<TOutputDb> {
     const selectQueryBuilder: SelectQueryBuilder<TModelDb> =
@@ -104,8 +113,16 @@ export class FindTypeOrmQueryBuilderService<
       >
     ).build(query, selectQueryBuilder);
 
-    return findByQueryBuilder(
-      findQueryTypeOrmOrQueryBuilder as SelectQueryBuilder<TModelDb>,
-    );
+    let outputDb: TOutputDb;
+
+    if (isQueryBuilder<TModelDb>(findQueryTypeOrmOrQueryBuilder)) {
+      outputDb = await findByQueryBuilder(
+        findQueryTypeOrmOrQueryBuilder as SelectQueryBuilder<TModelDb>,
+      );
+    } else {
+      outputDb = await findByFindManyOptions(findQueryTypeOrmOrQueryBuilder);
+    }
+
+    return outputDb;
   }
 }
