+++ packages/backend/libraries/backend-db/src/persistence/adapter/typeorm/utils/isQueryBuilder.ts
